+++ higan/sfc/coprocessor/armdsp/armdsp.hpp
@@ -15,8 +15,8 @@ struct ArmDSP : Processor::ARM, Coprocessor {
   auto bus_read(uint mode, uint32 addr) -> uint32 override;
   auto bus_write(uint mode, uint32 addr, uint32 word) -> void override;
 
-  auto mmio_read(uint addr, uint8 data) -> uint8;
-  auto mmio_write(uint addr, uint8 data) -> void;
+  auto mmio_read(uint24 addr, uint8 data) -> uint8;
+  auto mmio_write(uint24 addr, uint8 data) -> void;
 
   auto init() -> void;
   auto load() -> void;
