+++ higan/sfc/coprocessor/epsonrtc/epsonrtc.cpp
@@ -135,7 +135,7 @@ auto EpsonRTC::sync() -> void {
   resync = true;  //alert program that time has changed
 }
 
-auto EpsonRTC::read(uint addr, uint8 data) -> uint8 {
+auto EpsonRTC::read(uint24 addr, uint8 data) -> uint8 {
   cpu.synchronizeCoprocessors();
   addr &= 3;
 
@@ -160,7 +160,7 @@ auto EpsonRTC::read(uint addr, uint8 data) -> uint8 {
   return data;
 }
 
-auto EpsonRTC::write(uint addr, uint8 data) -> void {
+auto EpsonRTC::write(uint24 addr, uint8 data) -> void {
   cpu.synchronizeCoprocessors();
   addr &= 3, data &= 15;
