+++ higan/sfc/coprocessor/epsonrtc/epsonrtc.hpp
@@ -11,8 +11,8 @@ struct EpsonRTC : Coprocessor {
   auto reset() -> void;
   auto sync() -> void;
 
-  auto read(uint addr, uint8 data) -> uint8;
-  auto write(uint addr, uint8 data) -> void;
+  auto read(uint24 addr, uint8 data) -> uint8;
+  auto write(uint24 addr, uint8 data) -> void;
 
   auto serialize(serializer&) -> void;
