+++ higan/sfc/coprocessor/event/event.cpp
@@ -56,7 +56,7 @@ auto Event::reset() -> void {
   scoreSecondsRemaining = 0;
 }
 
-auto Event::mcuRead(uint addr, uint8 data) -> uint8 {
+auto Event::mcuRead(uint24 addr, uint8 data) -> uint8 {
   if(board == Board::CampusChallenge92) {
     uint id = 0;
     if(select == 0x09) id = 1;
@@ -92,17 +92,17 @@ auto Event::mcuRead(uint addr, uint8 data) -> uint8 {
   return data;
 }
 
-auto Event::mcuWrite(uint addr, uint8 data) -> void {
+auto Event::mcuWrite(uint24 addr, uint8 data) -> void {
 }
 
-auto Event::read(uint addr, uint8 data) -> uint8 {
+auto Event::read(uint24 addr, uint8 data) -> uint8 {
   if(addr == 0x106000 || addr == 0xc00000) {
     return status;
   }
   return data;
 }
 
-auto Event::write(uint addr, uint8 data) -> void {
+auto Event::write(uint24 addr, uint8 data) -> void {
   if(addr == 0x206000 || addr == 0xe00000) {
     select = data;
     if(timer && data == 0x09) {
