+++ higan/sfc/coprocessor/event/event.hpp
@@ -11,11 +11,11 @@ struct Event : Coprocessor {
   auto power() -> void;
   auto reset() -> void;
 
-  auto mcuRead(uint addr, uint8) -> uint8;
-  auto mcuWrite(uint addr, uint8) -> void;
+  auto mcuRead(uint24 addr, uint8) -> uint8;
+  auto mcuWrite(uint24 addr, uint8) -> void;
 
-  auto read(uint addr, uint8 data) -> uint8;
-  auto write(uint addr, uint8 data) -> void;
+  auto read(uint24 addr, uint8 data) -> uint8;
+  auto write(uint24 addr, uint8 data) -> void;
 
   auto serialize(serializer&) -> void;
