+++ higan/sfc/coprocessor/icd2/mmio/mmio.cpp
@@ -1,4 +1,4 @@
-auto ICD2::read(uint addr, uint8 data) -> uint8 {
+auto ICD2::read(uint24 addr, uint8 data) -> uint8 {
   addr &= 0x40ffff;
 
   //LY counter
@@ -38,7 +38,7 @@ auto ICD2::read(uint addr, uint8 data) -> uint8 {
   return 0x00;
 }
 
-auto ICD2::write(uint addr, uint8 data) -> void {
+auto ICD2::write(uint24 addr, uint8 data) -> void {
   addr &= 0xffff;
 
   //VRAM port
