+++ higan/sfc/coprocessor/mcc/mcc.hpp
@@ -10,14 +10,14 @@ struct MCC {
   auto power() -> void;
   auto reset() -> void;
 
-  auto memory_access(bool write, Memory& memory, uint addr, uint8 data) -> uint8;
-  auto mcu_access(bool write, uint addr, uint8 data) -> uint8;
+  auto memory_access(bool write, Memory& memory, uint24 addr, uint8 data) -> uint8;
+  auto mcu_access(bool write, uint24 addr, uint8 data) -> uint8;
 
-  auto mcu_read(uint addr, uint8 data) -> uint8;
-  auto mcu_write(uint addr, uint8 data) -> void;
+  auto mcu_read(uint24 addr, uint8 data) -> uint8;
+  auto mcu_write(uint24 addr, uint8 data) -> void;
 
-  auto read(uint addr, uint8 data) -> uint8;
-  auto write(uint addr, uint8 data) -> void;
+  auto read(uint24 addr, uint8 data) -> uint8;
+  auto write(uint24 addr, uint8 data) -> void;
 
   auto commit() -> void;
