+++ higan/sfc/coprocessor/msu1/msu1.hpp
@@ -10,8 +10,8 @@ struct MSU1 : Coprocessor {
   auto dataOpen() -> void;
   auto audioOpen() -> void;
 
-  auto mmioRead(uint addr, uint8 data) -> uint8;
-  auto mmioWrite(uint addr, uint8 data) -> void;
+  auto mmioRead(uint24 addr, uint8 data) -> uint8;
+  auto mmioWrite(uint24 addr, uint8 data) -> void;
 
   auto serialize(serializer&) -> void;
