+++ higan/sfc/coprocessor/nss/nss.cpp
@@ -23,11 +23,11 @@ auto NSS::set_dip(uint16 dip) -> void {
   this->dip = dip;
 }
 
-auto NSS::read(uint, uint8) -> uint8 {
+auto NSS::read(uint24, uint8) -> uint8 {
   return dip;
 }
 
-auto NSS::write(uint, uint8) -> void {
+auto NSS::write(uint24, uint8) -> void {
 }
 
 }
