+++ higan/sfc/coprocessor/obc1/obc1.hpp
@@ -5,8 +5,8 @@ struct OBC1 {
   auto power() -> void;
   auto reset() -> void;
 
-  auto read(uint addr, uint8 data) -> uint8;
-  auto write(uint addr, uint8 data) -> void;
+  auto read(uint24 addr, uint8 data) -> uint8;
+  auto write(uint24 addr, uint8 data) -> void;
 
   auto serialize(serializer&) -> void;
