+++ higan/sfc/coprocessor/sa1/bus/bus.hpp
@@ -1,12 +1,12 @@
 struct CPUIRAM : Memory {
   auto size() const -> uint;
-  alwaysinline auto read(uint, uint8 = 0) -> uint8;
-  alwaysinline auto write(uint, uint8) -> void;
+  alwaysinline auto read(uint24, uint8 = 0) -> uint8;
+  alwaysinline auto write(uint24, uint8) -> void;
 } cpuiram;
 
 struct CPUBWRAM : Memory {
   auto size() const -> uint;
-  alwaysinline auto read(uint, uint8 = 0) -> uint8;
-  alwaysinline auto write(uint, uint8) -> void;
+  alwaysinline auto read(uint24, uint8 = 0) -> uint8;
+  alwaysinline auto write(uint24, uint8) -> void;
   bool dma;
 } cpubwram;
